==> preprocessing/P2-process_index_md_RUC.ipynb <==
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
from geopandas import GeoDataFrame as gdf

## Import data for Multiple Deprivation Indices on the LSOA level

df_imd = gpd.read_file('data/Lower_Super_Output_Area_(LSOA)_IMD_2019__(OSGB1936).geojson')

df_imd.crs

df_imd.to_crs({'init': 'epsg:27700'}, inplace=True)

print(df_imd.crs)
df_imd.head()
df_imd.shape

df_imd['IMDRank'].nunique()

df_imd['IMDDecil'].nunique()

df_imd.head()

df_imd = df_imd[['lsoa11cd', 'IMDRank','IMDDecil']]
df_imd.head()

## Map LSOA data to MSOA

msoa_to_lsoa = pd.read_csv('data/Output_Area_to_LSOA_to_MSOA_to_Local_Authority_District_(December_2017)_Lookup_with_Area_Classifications_in_Great_Britain.csv')

msoa_to_lsoa = msoa_to_lsoa[['LSOA11CD','MSOA11CD']]
msoa_to_lsoa.drop_duplicates(inplace=True)
msoa_to_lsoa.shape

msoa_to_lsoa.drop(msoa_to_lsoa[~msoa_to_lsoa['LSOA11CD'].isin(df_imd['lsoa11cd'])].index,inplace=True)

msoa_to_lsoa.shape

msoa_imd = df_imd.merge(msoa_to_lsoa, left_on='lsoa11cd', right_on='LSOA11CD', how ='inner')
msoa_imd.head()

msoa1 = msoa_imd[['MSOA11CD', 'IMDRank']].groupby(['MSOA11CD']).mean()

msoa2 = msoa_imd[['MSOA11CD', 'IMDDecil']].groupby(['MSOA11CD']).median()

msoa_imd = msoa1.merge(msoa2, on=['MSOA11CD'])

#msoa_imd.to_csv('data/msoa_imd.csv')

msoa_imd.head()

## RUC Classification

df_ruc = pd.read_csv('data/Rural_Urban_Classification_(2011)_of_Lower_Layer_Super_Output_Areas_in_England_and_Wales.csv')

df_ruc = df_ruc[['LSOA11CD','RUC11']]

# merge with MSOA
msoa_ruc = df_ruc.merge(msoa_to_lsoa, on='LSOA11CD', how ='inner')
msoa_ruc.head()

msoa_ruc = msoa_ruc[['RUC11','MSOA11CD']].groupby(['MSOA11CD']).first()

#msoa_ruc.to_csv('data/msoa_ruc.csv')
