==> preprocessing/P3-process_household_elec_consumption.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

elec_2010 = pd.read_csv('raw_data_to_edit/LSOA_ELEC_2010.csv')

elec_2010.head()

elec_2010.shape

elec_2010.columns

elec_2010['Lower Layer Super Output Area (LSOA) Code'].nunique()

lsoa_codes_to_keep = pd.read_csv('data/lsoa_codes_reduced.csv', index_col=False)

lsoa_codes_to_keep.head()

df_to_reduce = elec_2010

df_to_reduce.columns

## Reduce data to LSOA codes in England and Wales

def func_to_reduce_LSOA_codes(df_to_reduce, lsoa_codes_to_keep, col_to_reduce, col_lsoa):
    """ Extract the rows of df_to_reduce that correspond to LSOA regions
    whose code are contained in lsoa_codes."""
    
    codes_to_keep = sorted(pd.Series(lsoa_codes_to_keep[col_lsoa]).astype('str'))
    codes_all = sorted(pd.Series(df_to_reduce[col_to_reduce]).astype('str'))
    codes_to_remove = list(set(codes_all) - set(codes_to_keep))
    return df_to_reduce[~df_to_reduce[col_to_reduce].isin(codes_to_remove)]

years = [2010,2011,2012,2013,2014,2015,2016,2017,2018,2019]

for year in years:
    df_to_reduce = pd.read_csv('raw_data_to_edit/LSOA_ELEC_{}.csv'.format(year))
    print(df_to_reduce.shape)
    print(df_to_reduce.columns)

col_to_reduce = 'Lower Layer Super Output Area (LSOA) Code'
col_lsoa = 'LSOA11CD'
for year in years:
    df_to_reduce = pd.read_csv('raw_data_to_edit/LSOA_ELEC_{}.csv'.format(year))
    df_reduced = func_to_reduce_LSOA_codes(df_to_reduce, lsoa_codes_to_keep, col_to_reduce, col_lsoa)
    df_reduced.to_csv('data/lsoa_elec_{}_reduced.csv'.format(year), index=False)

for year in years:
    test_df = pd.read_csv('data/lsoa_elec_{}_reduced.csv'.format(year))
    print(test_df.shape)

## MSOA granularity

Output dataframe with MSOA codes as rows and years as energy usage. For each year, reduce to MSOA and append to a master dataframe.

# Choose column to keep
col_name = 'Mean domestic electricity consumption \n(kWh per meter)'

year=2010
df = pd.read_csv('raw_data_to_edit/LSOA_ELEC_{}.csv'.format(year))
df.rename(columns={'Middle Layer Super Output Area (MSOA) Code':'MSOA11CD',
                   col_name:'energy_cons_kWh_{}'.format(year)},inplace=True)

df_msoa = pd.DataFrame(index=df.MSOA11CD)
for year in years:
    df_temp = pd.read_csv('raw_data_to_edit/LSOA_ELEC_{}.csv'.format(year))
    df_temp.rename(columns={'Middle Layer Super Output Area (MSOA) Code':'MSOA11CD',
                   col_name:'energy_cons_kWh_{}'.format(year)},inplace=True)
    df_msoa_year = df_temp[['MSOA11CD','energy_cons_kWh_{}'.format(year)]].groupby('MSOA11CD').sum()
    df_msoa = df_msoa.join(df_msoa_year)

df_msoa.drop_duplicates(inplace=True)

#df_msoa.to_csv('data/msoa_electricity_consumption_timeseries.csv')
