==> preprocessing/P4-process_charger_data.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

## Charger data 

There are two sources for charger data:
1. Grants funding private chargers provide a count of private chargers by region:  2014-2021 on the Local Authority (LA) level.
2. Public charger count by region: quarterly from 10/2019-04/2021 on LA level.

### Private chargers

column_names_private = ['LA_code','LA_name',
                '2015', '2016','2017', '2018','2019','2020','2021','Total']
df_private_chargers = pd.read_excel("raw_data_to_edit/electric-vehicle-charging-device-grant-statistics-april-2021.ods", 
                                   engine="odf", sheet_name ="EVG_02", skiprows=7, names=column_names_private)

print(df_private_chargers.shape)
print(df_private_chargers.columns)
df_private_chargers.head()

# First let's aggregate the summation over the years
for date_index in range(df_private_chargers.shape[1]-4):
    df_private_chargers.iloc[:,date_index+3] =df_private_chargers.iloc[:,date_index+2] + df_private_chargers.iloc[:,date_index+3]


df_private_chargers.head()

# Load look up table to map between MSOA and LSOA
region_lookup = pd.read_csv('data/Output_Area_to_LSOA_to_MSOA_to_Local_Authority_District_(December_2017)_Lookup_with_Area_Classifications_in_Great_Britain.csv')

print(region_lookup.shape)
print(region_lookup.columns)
la_unique =region_lookup.LAD17CD.unique()

# want to keep LA that we have the region lookup data for. Reduce the private charger data
df_private_chargers = df_private_chargers[df_private_chargers['LA_code'].isin(la_unique)]
print(df_private_chargers.shape)
print(len(la_unique))

# Note we have 22 missing LA regions. Fill these with zeros assuming there are no chargers here
missing_la_private = list(set(la_unique) - set(df_private_chargers.LA_code.unique()))

column_names_private = df_private_chargers.columns.values
column_names_private

# pad these missing LA regions into the dataframe assuming zero values
# add to the dataframes
df_missing_la_private = pd.DataFrame(columns=column_names_private)
for i in range(len(missing_la_private)):
    df_missing_la_private = df_missing_la_private.append(pd.Series({"LA_code": missing_la_private[i], "LA_name": missing_la_private[i],
                         column_names_private[2]: 0, column_names_private[3]: 0,
                         column_names_private[4]: 0, column_names_private[5]: 0,
                         column_names_private[6]: 0, column_names_private[7]: 0,
                         column_names_private[8]: 0, column_names_private[9]: 0}),ignore_index=True)

df_missing_la_private.shape

df_private_chargers= df_private_chargers.append(df_missing_la_private)
df_private_chargers.reset_index(drop=True)
print(df_private_chargers.shape)
print(df_private_chargers.LA_code.nunique())

## Convert to MSOA level

# for each LA we now have the value for chargers for each MSOA within that MSOA. 
# We can now merge with MSOA
df_private_chargers_msoa = pd.merge(df_private_chargers, region_lookup, right_on='LAD17CD',left_on='LA_code')

df_private_chargers_msoa.drop_duplicates(subset=['MSOA11CD'],inplace=True)
df_private_chargers_msoa = df_private_chargers_msoa[['MSOA11CD','LA_code',
                                                    '2015','2016','2017','2018','2019','2020','2021','Total']]
print(df_private_chargers_msoa.shape)
df_private_chargers_msoa.head()

df_private_chargers_msoa.MSOA11CD.nunique()

# Have lost some chargers
# count the number of MSOA in each LSOA
msoa_count = {}
for index, row in df_private_chargers.iterrows():
    msoa_count[row['LA_code']] = df_private_chargers_msoa[df_private_chargers_msoa['LA_code']==row['LA_code']].shape[0]

len(msoa_count)

la_to_msoa = pd.DataFrame.from_dict(msoa_count,orient='index',
                       columns=['msoa_count'])
la_to_msoa.reset_index(level=0, inplace=True)
la_to_msoa.head()

df_private_chargers_msoa = pd.merge(df_private_chargers_msoa, la_to_msoa,right_on='index',left_on='LA_code')
df_private_chargers_msoa.head()

df_private_chargers_msoa.columns

df_private_chargers_msoa['2015_msoa']=df_private_chargers_msoa['2015']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2016_msoa']=df_private_chargers_msoa['2016']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2017_msoa']=df_private_chargers_msoa['2017']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2018_msoa']=df_private_chargers_msoa['2018']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2019_msoa']=df_private_chargers_msoa['2019']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2020_msoa']=df_private_chargers_msoa['2020']/df_private_chargers_msoa['msoa_count']
df_private_chargers_msoa['2021_msoa']=df_private_chargers_msoa['2021']/df_private_chargers_msoa['msoa_count']

df_private_chargers_msoa.head()

print(df_private_chargers_msoa['2021_msoa'].sum())
print(df_private_chargers['2021'].sum())

df_private_chargers_msoa.shape

# drop unnecessary columns and save
df_private_chargers_msoa = df_private_chargers_msoa[['MSOA11CD','2015_msoa',
                                                     '2016_msoa','2017_msoa',
                                                     '2018_msoa','2019_msoa',
                                                     '2020_msoa','2021_msoa']]

#df_private_chargers_msoa.to_csv('data/private_chargers_msoa_timeseries.csv')

## Public chargers

column_names_public = ['LA_code','LA_name',
                '04-2021', 'Apr21_per100kpop',
               '01-2021', 'Jan21_per100kpop',
               '10-2020', 'Oct20_per100kpop',
               '07-2020', 'Jul20_per100kpop',
               '04-2020', 'Apr20_per100kpop',
               '01-2020', 'Jan20_per100kpop',
               '10-2019', 'Oct19_per100kpop']
df_public_chargers = pd.read_excel("raw_data_to_edit/electric-vehicle-charging-device-statistics-april-2021.ods", 
                                   engine="odf", sheet_name ="EVCD_01a", skiprows=7, names=column_names_public)

df_public_chargers = df_public_chargers[['LA_code','04-2021','01-2021',
                                         '10-2020','07-2020',
                                         '04-2020', '01-2020','10-2019']]
print(df_public_chargers.shape)
print(df_public_chargers.columns)
df_public_chargers.head()

df_public_chargers.LA_code.nunique()

# want to keep LA that we have the region lookup data for. Reduce the private charger data
df_public_chargers = df_public_chargers[df_public_chargers['LA_code'].isin(la_unique)]
print(df_public_chargers.shape)
print(len(la_unique))

# Again we have 22 missing regions

missing_la_public = list(set(la_unique) - set(df_public_chargers.LA_code.unique()))
column_names_public = df_public_chargers.columns.values
print(column_names_public)


df_missing_la_public = pd.DataFrame(columns=column_names_public)
for i in range(len(missing_la_public)):
    df_missing_la_public = df_missing_la_public.append(pd.Series({"LA_code": missing_la_public[i],
                         column_names_public[1]: 0, column_names_public[2]: 0,
                         column_names_public[3]: 0, column_names_public[4]: 0,
                         column_names_public[5]: 0, column_names_public[6]: 0, column_names_public[7]: 0}),ignore_index=True)

print(df_missing_la_public.shape)
df_public_chargers= df_public_chargers.append(df_missing_la_public)
df_public_chargers.reset_index(drop=True)
print(df_public_chargers.LA_code.nunique())
df_public_chargers.drop_duplicates(subset='LA_code',inplace=True)

print(df_public_chargers.shape)
df_public_chargers.head()

df_public_chargers_msoa = pd.merge(df_public_chargers, region_lookup, right_on='LAD17CD',left_on='LA_code')
df_public_chargers_msoa.drop_duplicates(subset=['MSOA11CD'],inplace=True)
df_public_chargers_msoa = df_public_chargers_msoa[['MSOA11CD','LA_code',
                                                    '10-2019','01-2020',
                                                   '04-2020','07-2020',
                                                   '10-2020','01-2021','04-2021']]
print(df_public_chargers_msoa.shape)
df_public_chargers_msoa.head()

# Aggregate by summation
# First let's aggregate the summation over the years
for date_index in range(df_public_chargers_msoa.shape[1]-3):
    df_public_chargers_msoa.iloc[:,date_index+3] =df_public_chargers_msoa.iloc[:,date_index+2] + df_public_chargers_msoa.iloc[:,date_index+3]


df_public_chargers_msoa.head()

df_public_chargers_msoa = pd.merge(df_public_chargers_msoa, la_to_msoa,right_on='index',left_on='LA_code')
df_public_chargers_msoa.head()

df_public_chargers_msoa['10-2019_msoa']=df_public_chargers_msoa['10-2019']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['01-2020_msoa']=df_public_chargers_msoa['01-2020']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['04-2020_msoa']=df_public_chargers_msoa['04-2020']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['07-2020_msoa']=df_public_chargers_msoa['07-2020']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['10-2020_msoa']=df_public_chargers_msoa['10-2020']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['01-2021_msoa']=df_public_chargers_msoa['01-2021']/df_public_chargers_msoa['msoa_count']
df_public_chargers_msoa['04-2021_msoa']=df_public_chargers_msoa['04-2021']/df_public_chargers_msoa['msoa_count']

df_public_chargers_msoa.head()

print(df_public_chargers_msoa['01-2020_msoa'].sum())
print(df_public_chargers['01-2020'].sum()+df_public_chargers['10-2019'].sum())

#df_public_chargers_msoa.to_csv('data/public_chargers_msoa_timeseries.csv')
