==> preprocessing/P5-process_pv_data.ipynb <==
import geopandas as gpd
import pandas as pd
import numpy as np

## Feed-in Tariff Installations

%%time 

df_fit_1 = pd.read_csv('../raw_data_to_edit/installation_report_apr2021_part_1.csv',skiprows=4)
df_fit_1.head()

df_fit_2 = pd.read_csv('../raw_data_to_edit/installation_report_apr2021_part_2.csv',skiprows=4)
df_fit_2.head()

df_fit_3 = pd.read_csv('../raw_data_to_edit/installation_report_apr2021_part_3.csv',skiprows=4)
df_fit_3.head()

df_fit = pd.concat([df_fit_1, df_fit_2, df_fit_3])

#only keep PV technology
df_fit = df_fit[df_fit['Technology'] == 'Photovoltaic']

df_fit.columns

df_fit['Installation Type'].unique()

#only keep domestic data
df_fit = df_fit[df_fit['Installation Type'] == 'Domestic']

df_fit.shape

Reduce to LSOA areas of interest

lsoa_codes_to_keep = pd.read_csv('../data/lsoa_codes_reduced.csv')
lsoa_codes_to_keep.head()

df_fit.head()

We need to reduce the full dataset to give values on the LSOA regions.

# make a new column of year/month data

print(df_fit['Application date '].min())
print(df_fit['Application date '].max())

df_fit['date_of_commission'] =  pd.to_datetime(df_fit['Commissioning date']).dt.to_period('M')
df_fit.head()

Create a pivot table of LSOA region by date counting the number of installations.

lsoa_date_pv_capacity = pd.pivot_table(df_fit, values='Installed capacity', index=['LLSOA Code'],
                    columns=['date_of_commission'], aggfunc=np.sum, fill_value=0)


lsoa_date_pv_count = pd.pivot_table(df_fit, values='Installed capacity', index=['LLSOA Code'],
                    columns=['date_of_commission'], aggfunc='count', fill_value=0)


lsoa_date_pv_count.describe()

lsoa_date_pv_count.head()

lsoa_date_pv_count.reset_index(level=0, inplace=True)
lsoa_date_pv_count.columns

lsoa_date_pv_count.rename(columns={'LLSOA Code': 'LSOA11CD'}, inplace=True)

lsoa_codes_to_keep = pd.read_csv('../data/lsoa_codes_reduced.csv')
lsoa_codes_to_keep.head()

df_fit_lsoa = pd.merge(
    lsoa_codes_to_keep,
    lsoa_date_pv_count,
    how="left",
    on='LSOA11CD',
)

df_fit_lsoa.shape

df_fit_lsoa.describe()

mean_across_years = df_fit_lsoa.mean()

# save to file
#df_fit_lsoa.to_csv('data/pv_lsoa_fit_timeseries.csv')

## Repeat for MSOA

df_fit_lsoa.shape

msoa_date_pv_count = pd.pivot_table(df_fit, values='Installed capacity', index=['MLSOA Code'],
                    columns=['date_of_commission'], aggfunc='count', fill_value=0)

msoa_date_pv_count.head()

print(msoa_date_pv_count.shape)
msoa_date_pv_count.describe()

# Throw away 2014 and earlier, aggregate by summation and save
msoa_date_pv_count.drop(msoa_date_pv_count.columns[:159],axis=1,inplace=True)

msoa_date_pv_count.head()

for date_index in range(msoa_date_pv_count.shape[1]-2):
    msoa_date_pv_count.iloc[:,date_index+2] = msoa_date_pv_count.iloc[:,date_index+1] + msoa_date_pv_count.iloc[:,date_index+2]

msoa_date_pv_count.describe()

#msoa_date_pv_count.to_csv('data/pv_msoa_fit_timeseries_2014.csv')
