==> preprocessing/P6-process_houseprices.ipynb <==
## Processing house price data

Imputing missing values.

import pandas as pd
import geopandas as gpd
import numpy as np

df_houseprice = pd.read_excel('data/houseprices_median_lsoa.xls', sheet_name ='Data', skiprows=5)

df_houseprice.head()

### Keep most recent data, and assume constant

df_houseprice = df_houseprice[['Local authority code', 'LSOA code', 'Year ending Dec 2019']]

## Function definition for converting int to float, and imputing srt to mean of float

def func_impute_to_float(df, col):
    str_mask = [type(x) == str for x in df[col]]
    int_mask = [type(x) == int for x in df[col]]
    float_mask = [type(x) == float for x in df[col]]
    type_mask = [type(x) for x in df[col]]
    
    if int in set(type_mask):
        #set to be float
        df[col][int_mask] = df[col][int_mask].astype('float')
        float_mask = [type(x) == float for x in df[col]]
    
    if str in set(type_mask):
        #set to be mean of numeric values
        df[col][str_mask] = df[col][float_mask].mean()
        
    df[col] = df[col].astype('float')
    
    assert df[col].dtypes == 'float64'
    
    return df

df_houseprice = func_impute_to_float(df_houseprice, 'Year ending Dec 2019')

# run over every column we want to be numeric
for col in df_houseprice.columns[4:]:
    df_houseprice = func_impute_to_float(df_houseprice, col)


df_houseprice.dtypes

# save as csv
#df_houseprice.to_csv('data/lsoa_houseprices.csv')

### Map to MSOA

msoa_to_lsoa = pd.read_csv('data/Output_Area_to_LSOA_to_MSOA_to_Local_Authority_District_(December_2017)_Lookup_with_Area_Classifications_in_Great_Britain.csv')

msoa_to_lsoa = msoa_to_lsoa[['LSOA11CD','MSOA11CD']]

print(msoa_to_lsoa.shape)
print(msoa_to_lsoa.LSOA11CD.nunique())
print(msoa_to_lsoa.MSOA11CD.nunique())

msoa_to_lsoa.drop_duplicates(inplace=True)
msoa_to_lsoa.shape

msoa_to_lsoa.LSOA11CD.nunique()

df_houseprice['LSOA code'].nunique()

# drop unnecessary lsoa regions
msoa_to_lsoa.drop(msoa_to_lsoa[~msoa_to_lsoa['LSOA11CD'].isin(df_houseprice['LSOA code'])].index,inplace=True)

msoa_to_lsoa.LSOA11CD.nunique()

# merge with df
msoa_houseprice = df_houseprice.merge(msoa_to_lsoa, left_on='LSOA code', right_on='LSOA11CD', how ='inner')

msoa_houseprice.head()

msoa_houseprice.MSOA11CD.nunique()

msoa_houseprice.shape

msoa_summed_houseprice = msoa_houseprice[['MSOA11CD', 'Year ending Dec 2019']].groupby(['MSOA11CD']).sum()

msoa_lsoa_count = msoa_houseprice[['MSOA11CD', 'LSOA11CD']].groupby(['MSOA11CD']).count()

msoa_avg_houseprice = msoa_summed_houseprice.merge(msoa_lsoa_count, on='MSOA11CD')
msoa_avg_houseprice['houseprice_avg_2019'] = msoa_avg_houseprice['Year ending Dec 2019'] / msoa_avg_houseprice['LSOA11CD']
msoa_avg_houseprice.head()

msoa_avg_houseprice.reset_index(inplace=True)
msoa_avg_houseprice = msoa_avg_houseprice[['MSOA11CD', 'houseprice_avg_2019']]
msoa_avg_houseprice.head()

#msoa_avg_houseprice.to_csv('data/msoa_houseprice.csv')
